fix: checks user/access sys
